+++ sdk/trace/config.go
@@ -16,7 +16,6 @@ package trace // import "go.opentelemetry.io/otel/sdk/trace"
 
 import (
 	"go.opentelemetry.io/otel/sdk/resource"
-	"go.opentelemetry.io/otel/sdk/trace/internal"
 )
 
 // Config represents the global tracing configuration.
@@ -25,7 +24,7 @@ type Config struct {
 	DefaultSampler Sampler
 
 	// IDGenerator is for internal use only.
-	IDGenerator internal.IDGenerator
+	IDGenerator IDGenerator
 
 	// MaxEventsPerSpan is max number of message events per span
 	MaxEventsPerSpan int
